add ntpdate-sync
The ntp daemon will only synchronize the clocks when the time
difference is not too big. It has command line option to allow to
set it initially, -g. Unfortanetly this does not seem to work if the
network is initially down. Furthermore running nptd -g -x and the
daemon at the same does not work either since they try to bind to the
same port. So use ntpdate-sync from a cron for now, which has the
disadvantage that the initial sync might take quite some time.
(and set some valid servers)